requirements.txt
langchain
langchain-huggingface
langchain_community
langchain_ollama
langchain_groq
faiss-cpu
python-dotenv
requests
flask
pypdf




#                     "Authorization": "eyJhbGciOiJBMTI4S1ciLCJlbmMiOiJBMTI4Q0JDLUhTMjU2In0.CgByx1MVUqjssxAW7IG0ymm6-qRpdPFtoEexYqcw22Gn2mLu3vIZ1A.2SjUidiSf0kKQAc4xwhaAg.Zh0_CJ8TaxiXm2n8avS-V0uOP0wMbcMuQc6Deb4zVKuxeL31NLbrTaFgjmxRDufMIE5MYZqyj9twJuOXJJUH5HJUH73GplG3F0-of8q8FY5bUwmwnhtiXYq4tkP_XIH1A9rtX46Ke_S2AOUJsrISVQinq8wbabKVM5o7zEsFCbRM9W2PHE_8M2pG8tD_q_A8.M0scFJRl1VjsJGUXrP2rgQ"  // Add the Bearer token for authorization

